Update with new options

+++ ArcWelderPlugin.py
@@ -1,5 +1,7 @@
 # Copyright (c) 2020 Aldo Hoeben / fieldOfView
 # The ArcWelderPlugin for Cura is released under the terms of the AGPLv3 or higher.
+# 25/11/2020 add options arcwelder_min_arc_segment / arcwelder_mm_per_arc_segment / arcwelder_allow_3d_arcs
+# 26/11/2020 change paramter : options min_arc_segment / mm_per_arc
 
 from collections import OrderedDict
 import json
@@ -33,7 +35,7 @@ class ArcWelderPlugin(Extension):
 
         ContainerRegistry.getInstance().containerLoadComplete.connect(self._onContainerLoadComplete)
         self._application.getOutputDeviceManager().writeStarted.connect(self._filterGcode)
-
+       
     def _onContainerLoadComplete(self, container_id: str) -> None:
         if not ContainerRegistry.getInstance().isLoaded(container_id):
             # skip containers that could not be loaded, or subsequent findContainers() will cause an infinite loop
@@ -94,7 +96,7 @@ class ArcWelderPlugin(Extension):
         global_container_stack = self._application.getGlobalContainerStack()
         if not global_container_stack:
             return
-
+           
         # get setting from Cura
         arcwelder_enable = global_container_stack.getProperty("arcwelder_enable", "value")
         if not arcwelder_enable:
@@ -112,7 +114,14 @@ class ArcWelderPlugin(Extension):
         maximum_radius = global_container_stack.getProperty("arcwelder_maximum_radius", "value")
         tolerance = global_container_stack.getProperty("arcwelder_tolerance", "value") / 100
         resolution = global_container_stack.getProperty("arcwelder_resolution", "value")
+        min_arc_segment = int(global_container_stack.getProperty("arcwelder_min_arc_segment", "value"))
+        mm_per_arc_segment = global_container_stack.getProperty("arcwelder_mm_per_arc_segment", "value")
+        allow_3d_arcs = global_container_stack.getProperty("arcwelder_allow_3d_arcs", "value")
 
+
+        # If the scene does not have a gcode, do nothing
+        if not hasattr(scene, "gcode_dict"):
+            return 
         gcode_dict = getattr(scene, "gcode_dict", {})
         if not gcode_dict: # this also checks for an empty dict
             Logger.log("w", "Scene has no gcode to process")
@@ -120,33 +129,46 @@ class ArcWelderPlugin(Extension):
 
         dict_changed = False
 
-        for plate_id in gcode_dict:
-            gcode_list = gcode_dict[plate_id]
-            if len(gcode_list) < 2:
-                Logger.log("w", "Plate %s does not contain any layers", plate_id)
-                continue
+        # get gcode list for the active build plate
+        active_build_plate_id = Application.getInstance().getMultiBuildPlateModel().activeBuildPlate
+        gcode_list = gcode_dict[active_build_plate_id]
+        if not gcode_list:
+            return
+            
+        if len(gcode_list) < 2:
+            Logger.log("w", "Plate %s does not contain any layers", plate_id)
+            return
 
-            if ";ARCWELDERPROCESSED\n" not in gcode_list[0]:
-                layer_separator = ";ARCWELDERPLUGIN_GCODELIST_SEPARATOR\n"
-                joined_gcode = layer_separator.join(gcode_list)
+        if ";ARCWELDERPROCESSED\n" not in gcode_list[0]:
+            layer_separator = ";ARCWELDERPLUGIN_GCODELIST_SEPARATOR\n"
+            joined_gcode = layer_separator.join(gcode_list)
 
-                file_descriptor, path = tempfile.mkstemp()
-                with os.fdopen(file_descriptor, 'w') as temporary_file:
-                    temporary_file.write(joined_gcode)
+            file_descriptor, path = tempfile.mkstemp()
+            with os.fdopen(file_descriptor, 'w') as temporary_file:
+                temporary_file.write(joined_gcode)
 
-                subprocess.run([arcwelder_path, "-m=%f" % maximum_radius, "-t=%f" % tolerance, "-r=%f" % resolution, path])
-
-                with open(path, "r") as temporary_file:
-                    result_gcode = temporary_file.read()
-                os.remove(path)
-
-                gcode_list = result_gcode.split(layer_separator)
-                gcode_list[0] += ";ARCWELDERPROCESSED\n"
-                gcode_dict[plate_id] = gcode_list
-                dict_changed = True
+            if min_arc_segment>0 :
+                if allow_3d_arcs :
+                    subprocess.run([arcwelder_path, "-z", "-s=%f" % mm_per_arc_segment, "-a=%d" % min_arc_segment, "-m=%f" % maximum_radius, "-t=%f" % tolerance, "-r=%f" % resolution , path])
+                else:
+                    subprocess.run([arcwelder_path, "-s=%f" % mm_per_arc_segment, "-a=%d" % min_arc_segment, "-m=%f" % maximum_radius, "-t=%f" % tolerance, "-r=%f" % resolution , path])
             else:
-                Logger.log("d", "Plate %s has already been processed", plate_id)
-                continue
+                if allow_3d_arcs :
+                    subprocess.run([arcwelder_path, "-z", "-m=%f" % maximum_radius, "-t=%f" % tolerance, "-r=%f" % resolution , path])     
+                else:
+                    subprocess.run([arcwelder_path, "-m=%f" % maximum_radius, "-t=%f" % tolerance, "-r=%f" % resolution , path])
+                
+            with open(path, "r") as temporary_file:
+                result_gcode = temporary_file.read()
+            os.remove(path)
+
+            gcode_list = result_gcode.split(layer_separator)
+            gcode_list[0] += ";ARCWELDERPROCESSED\n"
+            gcode_dict[active_build_plate_id] = gcode_list
+            dict_changed = True
+        else:
+            Logger.log("d", "Plate %s has already been processed", active_build_plate_id)
+            return
 
         if dict_changed:
             setattr(scene, "gcode_dict", gcode_dict)
